+++ apps/web/app/(dash)/home/homeVariants.ts
@@ -1,0 +1,50 @@
+export const variants = [
+	[
+		{
+			type: "text",
+			content: "Unlock your",
+		},
+		{
+			type: "highlighted",
+			content: " digital brain",
+		},
+	],
+	[
+		{
+			type: "text",
+			content: "Save",
+		},
+		{
+			type: "highlighted",
+			content: " everything.",
+		},
+		{
+			type: "text",
+			content: " Connect",
+		},
+		{
+			type: "highlighted",
+			content: " anything.",
+		},
+	],
+	[
+		{
+			type: "text",
+			content: "Turn your bookmarks into",
+		},
+		{
+			type: "highlighted",
+			content: " insights.",
+		},
+	],
+	[
+		{
+			type: "text",
+			content: "The smart way to use your",
+		},
+		{
+			type: "highlighted",
+			content: " digital treasure.",
+		},
+	],
+];
