+++ packages/tailwind-config/globals.css
@@ -234,10 +234,3 @@ body {
 	color: #369dfd;
 	background: #21303d;
 }
-
-.bg {
-	background-color: #171B1F;
-	opacity: 1;
-	background: radial-gradient(circle, transparent 20%, #171B1F 20%, #171B1F 80%, transparent 80%, transparent), radial-gradient(circle, transparent 20%, #171B1F 20%, #171B1F 80%, transparent 80%, transparent) 100px 100px, linear-gradient(#1a1f23 8px, transparent 8px) 0 -4px, linear-gradient(90deg, #1a1f23 8px, #171B1F 8px) -4px 0;
-	background-size: 100px 100px, 100px 100px, 100px 100px, 100px 100px;
-}
