+++ packages/ui/shadcn/command.tsx
@@ -37,8 +37,9 @@ const CommandDialog = ({ children, ...props }: CommandDialogProps) => {
 const CommandInput = React.forwardRef<
 	React.ElementRef<typeof CommandPrimitive.Input>,
 	React.ComponentPropsWithoutRef<typeof CommandPrimitive.Input>
->(({ className , ...props }, ref) => (
-	<div className="" cmdk-input-wrapper="">
+>(({ className, ...props }, ref) => (
+	<div className="flex items-center px-3" cmdk-input-wrapper="">
+		<Search className="mr-2 h-4 w-4 shrink-0 opacity-50" />
 		<CommandPrimitive.Input
 			ref={ref}
 			className={cn(
